+++ crates/macros/Cargo.toml
@@ -11,5 +11,4 @@ proc-macro = true
 syn = "1.0"
 quote = "1.0"
 proc-macro2 = "1.0"
-serde_json = "1.0"
 winmd = { path = "../winmd" }
