+++ crates/macros/src/lib.rs
@@ -1,10 +1,10 @@
 extern crate proc_macro;
 
 use proc_macro::{TokenStream, TokenTree};
-use winmd::{TypeLimits, TypeReader, TypeStage};
+use winmd::{dependencies, TypeLimits, TypeReader, TypeStage};
 
 use std::collections::BTreeSet;
-use std::path::{Path, PathBuf};
+use std::path::PathBuf;
 
 /// A macro for generating WinRT modules into the current module
 #[proc_macro]
@@ -152,7 +152,7 @@ fn parse_dependencies(
         let token = stream.peek();
         match token {
             Some(TokenTree::Literal(value)) => {
-                expand_paths(
+                dependencies::expand_paths(
                     value.to_string().trim_matches('"'),
                     &mut dependencies,
                     false,
@@ -166,7 +166,7 @@ fn parse_dependencies(
                 path.push(wind_dir_env);
                 path.push(SYSTEM32);
                 path.push("winmetadata");
-                expand_paths(path, &mut dependencies, false);
+                dependencies::expand_paths(path, &mut dependencies, false);
                 let _os = stream.next();
             }
             Some(TokenTree::Ident(value)) if value.to_string().as_str() == "nuget" => {
@@ -179,10 +179,7 @@ fn parse_dependencies(
                     },
                     "`nuget` must be followed by a `:`"
                 );
-                let mut path = workspace_root();
-
-                path.push("target");
-                path.push("nuget");
+                let mut path = winmd::dependencies::nuget_root();
 
                 let mut name = String::new();
                 while {
@@ -201,54 +198,12 @@ fn parse_dependencies(
                 }
                 path.push(name);
 
-                expand_paths(path, &mut dependencies, true);
+                dependencies::expand_paths(path, &mut dependencies, true);
             }
             _ => break,
         }
     }
     dependencies
-}
-
-/// Returns the paths to resolved dependencies
-fn expand_paths<P: AsRef<Path>>(dependency: P, result: &mut BTreeSet<PathBuf>, recurse: bool) {
-    let path = dependency.as_ref();
-
-    if path.is_dir() {
-        let paths = std::fs::read_dir(path).unwrap_or_else(|e| {
-            panic!(
-                "Could not read dependecy directory at path {:?}: {}",
-                path, e
-            )
-        });
-        for path in paths {
-            let path = path.expect("Could not read directory entry");
-            let path = path.path();
-            if path.is_file() && path.extension() == Some(std::ffi::OsStr::new("winmd")) {
-                result.insert(path);
-            } else if path.is_dir() && recurse {
-                expand_paths(path, result, recurse)
-            }
-        }
-    } else if path.is_file() && path.extension() == Some(std::ffi::OsStr::new("winmd")) {
-        result.insert(path.to_path_buf());
-    } else {
-        panic!("Dependency {} is not a file or directory", path.display());
-    }
-}
-
-fn workspace_root() -> PathBuf {
-    let output = std::process::Command::new("cargo")
-        .args(&["metadata"])
-        .output()
-        .expect("Could not run `cargo metadata`")
-        .stdout;
-    let value: serde_json::Map<String, serde_json::Value> =
-        serde_json::from_slice(&output).expect("`cargo metadata` did not return json.");
-    let path = match value.get("workspace_root") {
-        Some(serde_json::Value::String(s)) => s,
-        _ => panic!("`cargo metadata` json was not in expected format"),
-    };
-    PathBuf::from(path)
 }
 
 // Snake <-> camel casing is lossy so we go for character but not case conversion
