+++ examples/nuget-dependency/Cargo.toml
@@ -10,4 +10,7 @@ winrt = { path = "../.." }
 [build-dependencies]
 winrt = { path = "../.." }
 
+[package.metadata.nuget_dependencies]
+"Microsoft.Windows.SDK.Contracts" = "1.25.0"
+
 [workspace]
