+++ examples/nuget-dependency/build.rs
@@ -3,6 +3,7 @@ use winrt::Builder;
 fn main() {
     Builder::default()
         .insert_namespaces(&["windows.foundation"])
+        .insert_nuget(&["Microsoft.Windows.SDK.Contracts"])
         .build()
         .unwrap();
 }
